+++ adbnik/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.3.10"
+__version__ = "1.4.0"
 APP_TITLE = "Adbnik"